Move Character/interfaces to models/game/characters/interfaces
