+++ frontend/models/game/characters/Healer.php
@@ -12,7 +12,7 @@ namespace frontend\models\game\characters;
 use frontend\models\game\base\Bee;
 use frontend\models\game\base\BeeInterface;
 use frontend\models\game\base\HoneyPoolInterface;
-use frontend\models\game\interfaces\characters\HealerInterface;
+use frontend\models\game\characters\interfaces\HealerInterface;
 
 class Healer extends Bee implements HealerInterface
 {
