+++ frontend/models/game/characters/Player.php
@@ -11,7 +11,7 @@ namespace frontend\models\game\characters;
 
 use frontend\models\game\base\CharacterInterface;
 use frontend\models\game\GameInterface;
-use frontend\models\game\interfaces\characters\PlayerInterface;
+use frontend\models\game\characters\interfaces\PlayerInterface;
 
 class Player implements PlayerInterface
 {
