+++ frontend/models/game/characters/interfaces/PlayerInterface.php
@@ -6,7 +6,7 @@
  * Time: 03:12
  */
 
-namespace frontend\models\game\interfaces\characters;
+namespace frontend\models\game\characters\interfaces;
 
 
 use frontend\models\game\base\CharacterInterface;
