+++ tests/game/inside/CharacterPoolTest.php
@@ -14,7 +14,7 @@ use frontend\models\game\base\CharacterPool;
 use frontend\models\game\characters\Drone;
 use frontend\models\game\characters\Healer;
 use frontend\models\game\characters\Player;
-use frontend\models\game\interfaces\characters\PlayerInterface;
+use frontend\models\game\characters\interfaces\PlayerInterface;
 use frontend\models\game\characters\Queen;
 use frontend\models\game\characters\Worker;
 use frontend\models\game\Game;
